:recycle: move to usecases
